BigInteger should be mapped to NUMERIC/DECIMAL by default, not BIGINT!
I guess this probably changes some fairly ancient behavior, but c'mon,
the reason I'm using BigInteger is 'cos my numbers don't fit in Long!

+++ hibernate-core/src/main/java/org/hibernate/type/descriptor/java/internal/BigIntegerJavaDescriptor.java
@@ -30,7 +30,7 @@ public class BigIntegerJavaDescriptor extends AbstractNumericJavaDescriptor<BigI
 
 	@Override
 	public SqlTypeDescriptor getJdbcRecommendedSqlType(SqlTypeDescriptorIndicators context) {
-		return context.getTypeConfiguration().getSqlTypeDescriptorRegistry().getDescriptor( Types.BIGINT );
+		return context.getTypeConfiguration().getSqlTypeDescriptorRegistry().getDescriptor( Types.NUMERIC );
 	}
 
 	@Override
